+++ server/src/main/java/com/server/app/dto/request/OrderUpdateRequest.java
@@ -16,8 +16,8 @@ public class OrderUpdateRequest {
     private Short orderId;          // Güncellenecek Order’ın PK
 
     private String customerId;      // Customer FK
-    private Integer employeeId;     // Employee FK
-    private Short shipViaId;        // Shipper FK
+    private Long employeeId;     // Employee FK
+    private Long shipViaId;        // Shipper FK
     private LocalDate orderDate;
     private LocalDate requiredDate;
     private LocalDate shippedDate;
